Begin implementing the new track view in album mode
Be advised, it's not possible to escape a selection yet, so you'll
probably get stuck here if you test :)

We're going to use a simple layout, with a listbox to display the
various tracks. No doubt we'll need a new widget here to display
state for each MediaInfo (playing, etc.) and add some information
to the left hand side (Album+Artist+Year?).

WIP.