view: Enable kinetic scrolling
Kinetic scrolling allows touch screen users to flick using their finger,
instead of attempting to use the scrollbars.

Note: This is in no way an attempt to be mobile or tablet friendly.
Let's not have this conversation. Ever. It's because I have a Thinkpad
with a touchscreen, and I want this. <Explained />