Add BudgieMediaLabel skeleton
Instead of using a GtkLabel to represent tracks, etc, within
BudgieMediaView, we'll use a specialised class that can also store
a MediaInfo, represent the relevant information, and be queried
to enable playback with minimal fuss.

We'll also be able to do some nice rendering of the play status,
making use of 'dim-label', etc.