Beautify album art
Simply put, we take the existing album art image, and then make it look
like a real CD case, by compositing the image with a base CD image, the
scaled copy of the album art image, and finally a transparent overlay
to make it look slightly more realistic.

This will result in further changes to the loading mechanism, as this
kind of work really needs to be done in a background thread when we
start up Budgie.